% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arousalEnergy_fun.R
\name{arousalEnergy}
\alias{arousalEnergy}
\title{Calculate energy needed during an arousal from torpor to euthermia.}
\usage{
arousalEnergy(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{bat.params}{list of bat parameters output from \code{\link{batLoad}}}
}
\value{
returns enegry required to arouse from torpor to euthermia
}
\description{
\code{arousalEnergy} Calculate the enegry required to maintain
 eurthermic temperatures during an arousal from torpor.
}
\examples{
##CalcEnergyArousal Examples
data("mylu.params")

e <- -10:34 #temperature vector

r <- arousalEnergy(Ta = e, bat.params = mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Energy Expended",
     main = "arousalEnergy")
}
\references{
McKechnie & Wolf 2004 equation 3
}
\seealso{
\code{\link{coolEnergy}}, \code{\link{coolTime}},
 \code{\link{euthermicEnergry}},
\code{\link{flyingEnergy}}, \code{\link{flyingTime}}

Other Arousal Functions: \code{\link{arousalTime}},
  \code{\link{coolEnergy}}, \code{\link{euthermicEnergy}},
  \code{\link{flyingEnergy}}, \code{\link{flyingTime}}
}
\author{
Katie Haase
}
