% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arousalTime_fun.R
\name{arousalTime}
\alias{arousalTime}
\title{Calculate time to arouse from torpor to euthermic temperatures}
\usage{
arousalTime(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{bat.params}{list of parameters output from \code{\link{batLoad}}}
}
\description{
\code{arousalTime} Calculates time to arouse to euthermic conditions
given ambient conditions
}
\examples{
##CalcArousalTime
data("mylu.params")

e <- -10:34 #temperature vector

r <- arousalTime(Ta = e, bat.params = mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Time to Arrousal",
     main = "arousalTime")
}
\references{
McKechnie & Wolf 2004 equation 3
}
\seealso{
\code{\link{CalcEnergyArousal}}, \code{\link{CalcCoolTime}},
\code{\link{coolEnergy}}, \code{\link{euthermicEnergy}},
\code{\link{flyingEnergy}}, \code{\link{flyingTime}}

Other Arousal Functions: \code{\link{arousalEnergy}},
  \code{\link{coolEnergy}}, \code{\link{euthermicEnergy}},
  \code{\link{flyingEnergy}}, \code{\link{flyingTime}}
}
