% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batDynamic_fun.R
\name{batDynamic}
\alias{batDynamic}
\title{Deterministic Model of Fungal Growth}
\usage{
batDynamic(t, y, params)
}
\arguments{
\item{t}{time}

\item{y}{dependant varriables calculated within
\code{\link{hibernationModel}}}

\item{params}{parameters accociated with \code{bat.params},
and \code{fung.params}}
}
\description{
\code{DetModel} solves differential equasions accociated with the
deterministic model of fungal growth.
}
\details{
internal function for \code{\link{hibernationModel}} to calculate the
proportion of time and energy spent in the various states that compose a hibernation.
}
\examples{
## DynamicEnegryPd_ex
data("mylu.params")
data("fung.params")

fung.ch <- fungalSelect("Chaturvedi")


env.df <- buildEnv(temp = c(1:10),
                pct.rh = c(90:100),
                range.res.temp = 1,
                range.res.rh = 1,
                twinter = 9,
                winter.res = 4)

bingo <- hibernationModel(env.df, mylu.params, fung.ch)
}
\seealso{
\code{\link{hibernationModel}},  \code{\link{buildEnv}}

Other Model Engine: \code{\link{buildEnv}},
  \code{\link{hibernationModel}}
}
