% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatLoad_fun.R
\name{BatLoad}
\alias{BatLoad}
\title{Load static bat metabolic parameters}
\usage{
BatLoad(x, species)
}
\arguments{
\item{x}{dataframe from which parameters are taken see \strong{Details}}

\item{species}{character string for selected species from file above}
}
\value{
returns list of names parameters used for metabolic calculations
}
\description{
\code{BatLoad} Loads static metabolic parameters for bats from dataframe.
}
\details{
Function loads the metabolic parameters that will be used accross
the entire distribution.

\strong{RMR}: resting metabolic rate

\strong{TMR.min}: torpor metabolic rate

\strong{Teu}: temperature euthermic

\strong{Tlc}: lower critical temperature during euthermia

\strong{Ttormin}: temperature where \code{tmr.min} is achieved

\strong{Ceu}: euthermic thermal conductance

\strong{Ct}: thermal conductance during torpor

\strong{S}: specific heat capacity of tissue

\strong{k}: some constant

\strong{ttormax}: maximum time spent in torpor

\strong{tar}: time to arousal

\strong{teu}: time spent euthermic during arousals

\strong{mass}: mass of the critter
}
\examples{
##BatLoad Example
data("bat.params")
myyu.params <- BatLoad(bat.params, species = "M.californicus")
print(str(myyu.params))
}
