% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildEnv_fun.R
\name{BuildEnv}
\alias{BuildEnv}
\title{Build an environmental space to run the model across}
\usage{
BuildEnv(temp, hum, range.res = 1)
}
\arguments{
\item{temp}{a raster or a list of two numbers representing the minimum and maximum
temperatures to run the model across.}

\item{hum}{a raster or a list of two numebers representing the minimum and maximum
percent relative humidity to run the model across.}

\item{range.res}{a single numeber representing the resultion of
\code{temp.range} and \code{hum.range}.}
}
\value{
returns a dataframe containing all possiable combinations of
temperature and humidity conditions across which the model will be run.
}
\description{
\code{BuildEnv} builds the environmental parameters across which the
mechanistic model will be applied. This does not necessairly need to
represent real-world condtions but can instead be used to explore all of
parameter space.
}
\details{
This step can be a major determinate of how fast the model will run
. If the \code{temp.range} and \code{hum.range} and large or
\code{range.res} is especially fine, this may present computational
challenges downstream.
}
\examples{
##BuildEnv
m.env <- BuildEnv(temp = c(1,100), hum = c(1,100), range.res = 1)
}
