% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcArousalTime_fun.R
\name{CalcArousalTime}
\alias{CalcArousalTime}
\title{Calculate time to arouse from torpor to euthermic temperatures}
\usage{
CalcArousalTime(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{bat.params}{list of parameters output from \code{\link{BatLoad}}}
}
\description{
\code{CalcArousalTime} Calculates time to arouse to euthermic conditions
given ambient conditions
}
\examples{
##CalcArousalTime
data("mylu.params")

e <- -10:34 #temperature vector

r <- CalcArousalTime(Ta = e, bat.params = mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Time to Arrousal",
     main = "CalcArousalTime")
}
\references{
McKechnie & Wolf 2004 equation 3
}
\seealso{
\code{\link{CalcEnergyArousal}}, \code{\link{CalcCoolTime}},
\code{\link{CalcEnergyCool}}, \code{\link{CalcEnergyTimeEuthermic}}

Other Arousal Functions: \code{\link{CalcEnergyArousal}},
  \code{\link{CalcEnergyCool}},
  \code{\link{CalcEnergyTimeEuthermic}}
}
