% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEnergyArousal_fun.R
\name{CalcEnergyArousal}
\alias{CalcEnergyArousal}
\title{Calculate energy needed during an arousal from torpor to euthermia.}
\usage{
CalcEnergyArousal(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{bat.params}{list of bat parameters output from \code{\link{BatLoad}}}
}
\value{
returns enegry required to arouse from torpor to euthermia
}
\description{
\code{CalcEnergyArousal} Calculate the enegry required to maintain
 eurthermic temperatures during an arousal from torpor.
}
\examples{
##CalcEnergyArousal Examples
data("mylu.params")

e <- -10:34 #temperature vector

r <- CalcEnergyArousal(Ta = e, bat.params = mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Energy Expended",
     main = "CalcEnergyArousal")
}
\references{
McKechnie & Wolf 2004 equation 3
}
\seealso{
\code{\link{CalcEnergyCool}}, \code{\link{CalcCoolTime}},
\code{\link{CalcEnergyCool}}, \code{\link{CalcEnergyTimeEuthermic}},
\code{\link{calcEnergyFlying}}, \code{\link{CalcTimeFlying}}

Other Arousal Functions: \code{\link{CalcArousalTime}},
  \code{\link{CalcEnergyCool}},
  \code{\link{CalcEnergyFlying}},
  \code{\link{CalcEnergyTimeEuthermic}},
  \code{\link{CalcTimeFlying}}
}
\author{
Katie Haase
}
