% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEnergyCool_fun.R
\name{CalcEnergyCool}
\alias{CalcEnergyCool}
\title{Calculate energy needed during a cooling to torpor}
\usage{
CalcEnergyCool(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{bat.params}{list of bat parameters output from \code{\link{BatLoad}}}
}
\value{
returns energy required to cool
}
\description{
\code{CalcEnergyCool} Calculate the energy required to cool to torpid temperatures
}
\examples{
##CalcEnergyCool
data("mylu.params")

e <- -10:34 #temperature vector

r <- CalcEnergyCool(Ta = e, bat.params = mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Energy Expended",
     main = "CalcEnergyCool")
}
\references{
McKechnie & Wolf 2004 equation 3
}
\seealso{
\code{\link{CalcEnergyArousal}}, \code{\link{CalcCoolTime}},
\code{\link{CalcEnergyCool}}, \code{\link{CalcEnergyTimeEuthermic}}

Other Arousal Functions: \code{\link{CalcArousalTime}},
  \code{\link{CalcEnergyArousal}},
  \code{\link{CalcEnergyTimeEuthermic}}
}
\author{
Katie Haase
}
