% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEnergyFlying.R
\name{CalcEnergyFlying}
\alias{CalcEnergyFlying}
\title{Calculate the energy expended during flight}
\usage{
CalcEnergyFlying(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient tempterature}

\item{bat.params}{see \code{\link{BatLoad}}}
}
\value{
energy expended during a time unit of flight
}
\description{
\code{CalcEnergyFlying} calculates energry expended during a unit of light
}
