% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEnergyTimeEuthermic_fun.R
\name{CalcEnergyTimeEuthermic}
\alias{CalcEnergyTimeEuthermic}
\title{Calculate energy per time euthermic with ambient temperature}
\usage{
CalcEnergyTimeEuthermic(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{bat.params}{list of bat parameters output from \code{\link{BatLoad}}}
}
\value{
TODO find units accociated with it
}
\description{
\code{CalcEnergyTimeEuthermic} Function calculating the energy required to
maintain euthermia for a given time and ambient temperature.
}
\examples{
##CalEnergyTimeEuthermic Example
data("mylu.params")

e <- -10:34 #temperature vector

r <- CalcEnergyTimeEuthermic(e, mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Energy Expended",
     main = "CalEnergyTimeEuthermic")
}
\seealso{
\code{\link{CalcEnergyCool}}, \code{\link{CalcCoolTime}},
\code{\link{CalcEnergyCool}}, \code{\link{CalcEnergyCool}},
\code{\link{calcEnergyFlying}}, \code{\link{CalcTimeFlying}}

Other Arousal Functions: \code{\link{CalcArousalTime}},
  \code{\link{CalcEnergyArousal}},
  \code{\link{CalcEnergyCool}},
  \code{\link{CalcEnergyFlying}},
  \code{\link{CalcTimeFlying}}
}
