% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEnergyTimeTorpid_fun.R
\name{CalcEnergyTimeTorpid}
\alias{CalcEnergyTimeTorpid}
\title{Calculate energy per time at toprid body temp given ambient temperature}
\usage{
CalcEnergyTimeTorpid(Ta, bat.params, q = CalcQ(Ta))
}
\arguments{
\item{Ta}{ambient emperature}

\item{bat.params}{list of bat parameters output from \code{\link{BatLoad}}}

\item{q}{Q10 scaling value}
}
\value{
TODO find units accociated with
}
\description{
\code{CalcEnergyTimeTorpid} Function calculating the enegry required for a
bout of torpor at a ambient temperature
}
\examples{
##CalcEnergyTimeTorpid Example
data("mylu.params")

e <- -10:34 #temperature vector

r <- CalcEnergyTimeTorpid(e, mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Energy Expended",
     main = "CalcEnergyTimeTorpid")
}
\author{
Katie Haase
}
