% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcQ_fun.R
\name{CalcQ}
\alias{CalcQ}
\title{Q10 scaling value as a function of ambient temperature}
\usage{
CalcQ(Ta, Q1 = 1.6, Q2 = 0.26, Q3 = 0.006)
}
\arguments{
\item{Ta}{ambient temperature in degrees C}

\item{Q1}{Q constant 1 = 1.6}

\item{Q2}{Q constant 2 = 0.26}

\item{Q3}{Q constant 3 = 0.006}
}
\value{
Returns the value of Q10 value for the given ambient temperature
}
\description{
\code{CalcQ} Scales metabolic rate with ambient temperature
}
\examples{
CalcQ(4)
CalcQ(6, 1.6, 0.26, 0.006)
}
