% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTimeEuthermic_R.R
\name{CalcTimeEuthermic}
\alias{CalcTimeEuthermic}
\title{Calculate time inactive during euthermia}
\usage{
CalcTimeEuthermic(bat.params)
}
\arguments{
\item{bat.params}{see \code{\link{BatLoad}}}
}
\value{
returns the time spent in euthermia (but not flying)
}
\description{
\code{CalcTimeEuthermic}
}
