% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTorporTime_fun.R
\name{CalcTorporTime}
\alias{CalcTorporTime}
\title{Calculate maximum time in torpor}
\usage{
CalcTorporTime(Ta, areaPd, inf, bat.params, q = CalcQ(Ta))
}
\arguments{
\item{Ta}{ambient temperature}

\item{areaPd}{area of pd}

\item{inf}{infection status (TRUE or FALSE)}

\item{bat.params}{list of bat parameters output from \code{\link{BatLoad}}}

\item{q}{TODO}
}
\value{
returns time bat spends in torpor
}
\description{
\code{CalcTorporTime} Calculate time of a torpor bout given ambient
temperature and infection status
}
\examples{
##CalcTorporTIme Example
data("mylu.params")

e <- -10:34 #temperature vector
a <- seq(from = .001, to = 3, length.out = 45) #area vector

CalcTorporTime(Ta = 4, areaPd = 0, inf = TRUE, bat.params = mylu.params)

r  <- CalcTorporTime(Ta = e, areaPd = rep(0,45), inf = FALSE, bat.params = mylu.params)
r1 <- CalcTorporTime(Ta = rep(4,45), areaPd = a, inf = TRUE, bat.params = mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Time in Torpor",
     main = "CalcTorporTIme")
plot(a, r1, xlab = "Area Infected (cm2^)", ylab = "Time in Torpor",
     main = "CalcTorporTIme")
}
