% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolEnergy_fun.R
\name{coolEnergy}
\alias{coolEnergy}
\title{Calculate energy needed during a cooling to torpor}
\usage{
coolEnergy(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{bat.params}{list of bat parameters output from \code{\link{batLoad}}}
}
\value{
returns energy required to cool
}
\description{
\code{coolEnergy} Calculate the energy required to cool to torpid temperatures
}
\examples{
##CalcEnergyCool
data("mylu.params")

e <- -10:34 #temperature vector

r <- coolEnergy(Ta = e, bat.params = mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Energy Expended",
     main = "coolEnergy")
}
\references{
McKechnie & Wolf 2004 equation 3
}
\seealso{
\code{\link{arousalEnergy}}, \code{\link{coolTime}},
\code{\link{coolEnergy}}, \code{\link{euthermicEnergy}},
\code{\link{flyingEnergy}}, \code{\link{flyingTime}}

Other Arousal Functions: \code{\link{arousalEnergy}},
  \code{\link{arousalTime}}, \code{\link{euthermicEnergy}},
  \code{\link{flyingEnergy}}, \code{\link{flyingTime}}
}
\author{
Katie Haase
}
