% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolTime_fun.R
\name{coolTime}
\alias{coolTime}
\title{Calculate time to cool from euthermic to torpor.}
\usage{
coolTime(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{bat.params}{list of parameters output from \code{\link{batLoad}}}
}
\value{
Time in hours.
}
\description{
\code{coolTime} Calculates time to cool from euthermic conditions to
torpid state given the ambient temperature.
}
\examples{
##CalcCoolTime
data("mylu.params")

e <- seq(-10:34) #temperature vector

r <- coolTime(Ta = e, bat.params = mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Time")
}
\references{
McKechnie & Wolf 2004 equation 3
}
\author{
Katie Haase
}
