% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DangerZone_fun.R
\name{dangerZone}
\alias{dangerZone}
\title{Plotting the difference between infected and and uninfected hibernating
bats.}
\usage{
dangerZone(mod.df, title, save.name = NULL, ...)
}
\arguments{
\item{mod.df}{results data from \code{\link{hibernationModel}}}

\item{title}{title for the figure}

\item{save.name}{a name and file path relative to the working directory
where the figure will be saved.}

\item{...}{used to pass arguments to \code{\link[ggplot2]{ggsave}}}
}
\value{
returns an environmental surface accross which the difference
between infected and uninfected hibernation times is plotted.
}
\description{
\code{dangerZone} plots the difference in hibernation length potentials
between infected and uninfected bats accross the environmental space.
}
\details{
Function will be updated in the future to include more plotting
parameter accessability, and potentially a raster method as well.
\url{https://www.youtube.com/watch?v=kyAn3fSs8_A}
}
\seealso{
\code{\link{survivalMultiplot}}; \code{\link{survivalPlotter}}; \code{\link{survivalHistogram}}

Other Plot Functions: \code{\link{survivalMultiplot}},
  \code{\link{survivalPlotter}},
  \code{\link{survivalRaster}}
}
\author{
L. Cain, K. Loggins
}
