% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoversionFunctions.R
\name{day.to.hour}
\alias{day.to.hour}
\title{\code{day.to.hour}}
\usage{
day.to.hour(x)
}
\arguments{
\item{x}{days to convert}
}
\description{
\code{day.to.hour}
}
\details{
general functions for consistant exchange between time units
}
