% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coversionFunctions.R
\name{day.to.month}
\alias{day.to.month}
\title{\code{day.to.month}}
\usage{
day.to.month(x)
}
\arguments{
\item{x}{input number of days}
}
\description{
\code{day.to.month}
}
\details{
general functions for consistant exchange between time units
}
