% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetModel_fun.R
\name{DetModel}
\alias{DetModel}
\title{Deterministic Model of Fungal Growth}
\usage{
DetModel(t, y, params)
}
\arguments{
\item{t}{tmeperature I guess? does it even get used? TODO}

\item{y}{dependant varriables calculated within
\code{\link{DynamicEnergyPd}}}

\item{params}{parameters passed through \code{\link{DynamicEnergyPd}}}
}
\description{
\code{DetModel} solves differential equasions accociated with the
deterministic model of fungal growth.
}
\details{
This may have room for improvement in it and may change in the
future
}
\examples{
## DynamicEnegryPd_ex
data("bat.params")
data("fung.params")

my.params <- bat.params["M.californicus",]
fung.ch <- FungSelect("Chaturvedi")


env.df <- BuildEnv(temp = c(1:10),
                hum = c(90:100),
                range.res = 1)

twinter <- seq(from = 0, to = 9*24*30*1, by = 24)
bingo <- DynamicEnergyPd(env.df, my.params, fung.ch)
}
\seealso{
\code{\link{DynamicEnergyPd}}
}
