% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffHist_fun.R
\name{DiffHist}
\alias{DiffHist}
\title{Creats histograms displaying the model resulst from the species distribution}
\usage{
DiffHist(surv.stk, dist.map, species.name, nights, save.name = NULL, ...)
}
\arguments{
\item{surv.stk}{raster result stack from \code{\link{SurvivalRaster}}}

\item{dist.map}{shapefile distribution of the species being modeled}

\item{species.name}{name of the species for plotting (generally a 4 letter
abreviation will work best)}

\item{nights}{a raster layer reprensenting the length of winter measured in nigths}

\item{save.name}{optional argument. When a path is passed files will save to that
location with provided name. If left as defult 'NULL' will print to plot window}

\item{...}{more things that are generally usefully but yeah.}
}
\value{
returns two histograms overlaying one another with median estimations
}
\description{
\code{DiffHist} Two hisograms are created and overlayed over one another.
The histograms represent the survival capasitites obtained through
\code{\link{DynamicEnergyPd}} across the distribution. one will contain the
results of null hibernation, and the other infection with WNS.
}
\details{
This function will likely go the way of the dodo with the next model
updates and build because this is unnecessairy convaluted.
}
\seealso{
\code{\link{DangerZone}}; \code{\link{MapFigs}}; \code{\link{SurvPlotter}}

Other PlotFunctions: \code{\link{DangerZone}},
  \code{\link{MapFigs}}, \code{\link{SurvPlotter}},
  \code{\link{SurvivalRaster}}
}
