% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicEnergyPd_fun.R
\name{DynamicEnergyPd}
\alias{DynamicEnergyPd}
\title{Dynamic energy cost calculation}
\usage{
DynamicEnergyPd(env.df, bat.params, fung.params)
}
\arguments{
\item{env.df}{dataframe conataing range of env conditions}

\item{bat.params}{parameters returned by \code{\link[pkg:batwintor]{BatLoad}}}

\item{fung.params}{parameters returned by \code{\link[pkg:batwintor]{FungLoad}})}
}
\description{
\code{DynamicEnergyPd} determines the growth area of Pd and the sums the
energy consumed under specified environmental conditions.
}
\details{
TODO
}
\examples{
## DynamicEnegryPd_ex
data("bat.params")
data("fung.params")

my.params <- bat.params["M.californicus",]
fung.ch <- FungSelect("Chaturvedi")


env.df <- BuildEnv(temp = c(1:10),
                hum = c(90:100),
                range.res = 1)

twinter <- seq(from = 0, to = 9*24*30*1, by = 24)
bingo <- DynamicEnergyPd(env.df, my.params, fung.ch)
}
\seealso{
\code{\link[pkg:batwintor]{DetModel}}
}
