% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnergyPD_fun.R
\name{EnergyPd}
\alias{EnergyPd}
\title{Calculate energy use and fungal area as a function of environmental
condtions and infection status}
\usage{
EnergyPd(Ta, twinter, Hd, WNS, bat.params, fung.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{twinter}{length of winter}

\item{Hd}{humidity}

\item{WNS}{logical, should the bats have WNS?}

\item{bat.params}{parameters returned by \code{\link{BatLoad}}}

\item{fung.params}{parameters returned by \code{\link{FungLoad}}}
}
\description{
\code{EnergyPd} calculates energy consumed and the area of funal growth
given static environmental conditions across a predefined length of winter.
}
