% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euthermicTime_fun.R.R
\name{euthermicTime}
\alias{euthermicTime}
\title{Calculate time inactive during euthermia}
\usage{
euthermicTime(bat.params)
}
\arguments{
\item{bat.params}{see \code{\link{batLoad}}}
}
\value{
returns the time spent in euthermia (but not flying)
}
\description{
\code{euthermicTime}
}
