% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewl_fun.R
\name{ewl}
\alias{ewl}
\title{Calculate Evaporative Water Loss}
\usage{
ewl(Ta, pct.rh, t, areaPd, fung.params, bat.params, torpid = TRUE,
  WNS = c(TRUE, FALSE))
}
\arguments{
\item{Ta}{ambient temperature (degrees C)}

\item{pct.rh}{precent relative humidity}

\item{t}{time}

\item{areaPd}{area of fungal growth (cm2)}

\item{fung.params}{parameters returned by \code{fung.params}}

\item{bat.params}{parameters returned by \code{bat.params}}

\item{torpid}{logical, if the animal is storpid or not}

\item{WNS}{logical, if the animal is infected with Pd or not}
}
\description{
\code{ewl} calculates evaporative waterloss based on species specific parameter.
}
\details{
The function includes mechanisms to deal with the evaporative water loss (EWL)
of both torpid and euthermic individuals, as well as increasing metabolic rates
for those infected with Pd.
}
\seealso{
\code{\link{torporEnergy}}

Other Torpor Functions: \code{\link{torporEnergy}}
}
\author{
Katie Haase
}
