% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyingEnergy_fun.R
\name{flyingEnergy}
\alias{flyingEnergy}
\title{Calculate the energy expended during flight}
\usage{
flyingEnergy(Ta, bat.params)
}
\arguments{
\item{Ta}{ambient tempterature}

\item{bat.params}{see \code{\link{batLoad}}}
}
\value{
energy expended during a time unit of flight
}
\description{
\code{flyingEnergy} calculates energry expended during a unit of light
}
\seealso{
\code{\link{CalcEnergyArousal}}, \code{\link{CalcCoolTime}},
\code{\link{coolEnergy}}, \code{\link{euthermicTime}},
\code{\link{CalcEnergyArousal}}, \code{\link{euthermicEnergy}},
\code{\link{flyingEnergy}}, \code{\link{flyingTime}}

Other Arousal Functions: \code{\link{arousalEnergy}},
  \code{\link{arousalTime}}, \code{\link{coolEnergy}},
  \code{\link{euthermicEnergy}}, \code{\link{flyingTime}}
}
