% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flyingTime_fun.R
\name{flyingTime}
\alias{flyingTime}
\title{Calculate the time spent flying}
\usage{
flyingTime(bat.params)
}
\arguments{
\item{bat.params}{see \code{\link{batLoad}}}
}
\value{
a time
}
\description{
\code{flyingTime} calculates the time spent flying as a proportion of
time in euthermic arrousals
}
\seealso{
\code{\link{CalcEnergyArousal}}, \code{\link{CalcCoolTime}},
\code{\link{coolEnergy}}, \code{\link{euthermicEnergy}},
\code{\link{flyingEnergy}}

Other Arousal Functions: \code{\link{arousalEnergy}},
  \code{\link{arousalTime}}, \code{\link{coolEnergy}},
  \code{\link{euthermicEnergy}}, \code{\link{flyingEnergy}}
}
