% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FungDat.R
\docType{data}
\name{fung.params}
\alias{fung.params}
\title{Growth coeffecients for funal growth models}
\format{a data set including 3 beta parameters and 2 mu parameters for 2
 models
\describe{
  \item{beta1}{beta1}
  \item{beta2}{beta2}
  \item{beta3}{beta3}
  \item{mu1}{mu1}
  \item{mu2}{mu2}
}}
\usage{
fung.params
}
\description{
The \eqn{\beta} and \eqn{\mu} growth coeffecients for two different models of fungal growth.
}
\details{
fungal growth equations are fit to Michaelis Menten equations that
reach maximal values at saturation.
}
\references{
See Haymen et al. 2016 for fitting details
}
\seealso{
Other Data: \code{\link{bat.params}}
}
\keyword{datasets}
