% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungalGrowth_fun.R
\name{fungalGrowth}
\alias{fungalGrowth}
\title{How fungal growth rate varries with body temperature}
\usage{
fungalGrowth(Tb, fung.params, t.min = 0)
}
\arguments{
\item{Tb}{body temperature of the animal in degrees C}

\item{fung.params}{list of parameters passed through TODO NameFunction}

\item{t.min}{temperature minimum (= 0 degrees C)}
}
\value{
returns temperature determinate growth rate
}
\description{
\code{fungalGrowth} Scales chosen fungal growth rate with to the
body temperautre of the infected animal.
}
\examples{
##FungalGrowthRate
data("fung.params")
fung.param <- fungalSelect("Chaturvedi")

e <- -10:20

r <- fungalGrowth(e, fung.params = fung.param)

plot(e, r, xlab = "Temp (C)", ylab = "Growth rate",
     main = "Fungal Growth Rate")
}
\seealso{
\code{\link{fungalLoad}},  \code{\link{fungalSelect}},
\code{\link{scaleFungalGrowth}}
}
