% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FungalGrowthRate_fun.R
\name{FungalGrowthRate}
\alias{FungalGrowthRate}
\title{How fungal growth rate varries with body temperature}
\usage{
FungalGrowthRate(Tb, fung.params, t.min = 0)
}
\arguments{
\item{Tb}{body temperature of the animal in degrees C}

\item{fung.params}{list of parameters passed through TODO NameFunction}

\item{t.min}{temperature minimum (= 0 degrees C)}
}
\value{
returns temperature determinate growth rate
}
\description{
\code{FungalGrowthRate} Scales chosen fungal growth rate with to the
body temperautre of the infected animal.
}
\examples{
##FungalGrowthRate
data("fung.params")
fung.param <- FungSelect("Chaturvedi")

e <- -10:20

r <- FungalGrowthRate(e, fung.params = fung.param)

plot(e, r, xlab = "Temp (C)", ylab = "Growth rate",
     main = "Fungal Growth Rate")
}
\seealso{
\code{\link{FungLoad}}
}
