% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fungalLoad_fun.R
\name{fungalLoad}
\alias{fungalLoad}
\title{Load parameter sets for fungal growth}
\usage{
fungalLoad(path.to.params, growth.option)
}
\arguments{
\item{path.to.params}{file path to the parameter files (without trailing "/")}

\item{growth.option}{fungal growth parameters see \strong{Details}}
}
\value{
Returns a named list of fungal growth scaling parameters
}
\description{
\code{fungalLoad} Loads fungal growth parameters from file and sets to selected option.
}
\details{
Growth parameters are either: "Chaturvedi" (faster growth) from
chatruvedi et al. PLoS One, ot "Verant" (slower growth rate), from Verent et
al. PLoS One.

\strong{mu1}: scaling parameter for Michaelis-Menton function

\strong{mu2}: scaling parameter for Michaelis-Menton function

\strong{beta1}: temperature dependant hourly rate shape parameter

\strong{beta2}: temperature dependant hourly rate shape parameter

\strong{beta3}: temperature dependant hourly rate shape parameter
}
\note{
Requires both files to be in same directory with names
\code{rate.parms.csv} and \code{humid.parms.csv}. These files are
shipped with this package, however if you would like to alter these
paremeters you may do so through their associated \code{.csv} files.
}
\examples{
##FungLoad
foo <- "paramFiles"
fung.param <- fungalLoad(path.to.params = foo,
                       growth.option = "Verant")
}
\seealso{
\code{\link{fungalGrowth}},  \code{\link{fungalSelect}},
\code{\link{scaleFungalGrowth}}
}
