% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDist_fun.R
\name{GetDist}
\alias{GetDist}
\title{Obtatin the species distribution for a species.}
\usage{
GetDist(path.to.data, species)
}
\arguments{
\item{path.to.data}{a path to the IUCN MAMMTERR data source.}

\item{species}{binomial latin name (or list of names) as a character
string.}
}
\value{
Returns a named shapefile(s) of the speceies input to the global
environment
}
\description{
\code{GetDist} obtain the IUCN distribution for a mammaian species.
}
\details{
The IUCN MAMMTERR data set includes distribution shapefiles for
most terrestial mammals however they are not without their flaws so use
accordingly. Data can be sourced from
\url{http://www.iucnredlist.org/technical-documents/spatial-data}
}
\note{
This may take a while as the IUCN database is large. May be a point
to fix in the future.

#@example TODO? is it needed
}
