% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hibernationModel_fun.R
\name{hibernationModel}
\alias{hibernationModel}
\title{Dynamic energy cost calculation}
\usage{
hibernationModel(env, bat.params, fung.params)
}
\arguments{
\item{env}{an object produced by \code{\link{buildEnv}} conataing range of env conditions
and a time vector to run the model across}

\item{bat.params}{parameters returned by \code{bat.params}}

\item{fung.params}{parameters returned by \code{fung.params}}
}
\value{
returns a dataframe containing a model results for each set of environmental
conditions at each time point.
}
\description{
\code{hibernationModel} determines the growth area of Pd and the sums the
energy consumed under specified environmental conditions.
}
\details{
This function formats, data going into, and processes data coming out of
\code{\link{batDynamic}}, and is the main main function of the package.
Outputs from ths function can be exceedingly large, and take a heckin' long time
so please plan accordingly.
}
\examples{
## DynamicEnegryPd_ex
data("mylu.params")
data("fung.params")

fung.ch <- fungalSelect("Chaturvedi")


env.df <- buildEnv(temp = c(1:10),
                pct.rh = c(90:100),
                range.res.temp = 1,
                range.res.rh = 1,
                twinter = 9,
                winter.res = 4)

bingo <- hibernationModel(env.df, mylu.params, fung.ch)
}
\seealso{
\code{\link{batDynamic}}, \code{\link{buildEnv}}

Other Model Engine: \code{\link{batDynamic}},
  \code{\link{buildEnv}}
}
