% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coversionFunctions.R
\name{hour.to.day}
\alias{hour.to.day}
\title{\code{hour.to.day}}
\usage{
hour.to.day(x)
}
\arguments{
\item{x}{input number of hours}
}
\description{
\code{hour.to.day}
}
\details{
general functions for consistant exchange between time units
}
