% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coversionFunctions.R
\name{hour.to.month}
\alias{hour.to.month}
\title{conversion functions}
\usage{
hour.to.month(x)
}
\arguments{
\item{x}{input value of hours}
}
\description{
\code{hour.to.month}
}
\details{
general functions for consistant exchange between time units
}
