% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microSpace_fun.R
\name{microSpace}
\alias{microSpace}
\title{microSpace}
\usage{
microSpace(mod.df, save.name = NULL, ...)
}
\arguments{
\item{mod.df}{Model dataframe returned from \code{\link{hibernationModel}}}

\item{save.name}{optional agrument for saving}

\item{...}{agruments to be passed to \code{\link{ggsave}}}
}
\description{
\code{microSpace} displays the influence of micro habitat selection
on predicted bat hibernation length.
}
