% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coversionFunctions.R
\name{month.to.hour}
\alias{month.to.hour}
\title{\code{month.to.hour}}
\usage{
month.to.hour(x)
}
\arguments{
\item{x}{input value of months}
}
\description{
\code{month.to.hour}
}
\details{
general functions for consistant exchange between time units
}
