% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrRaw_fun.R
\name{mrRaw}
\alias{mrRaw}
\title{Takes raw metabolic data and returns cleaned estimates for each species.}
\usage{
mrRaw(x, species.option)
}
\arguments{
\item{x}{a dataframe}

\item{species.option}{species name or code (as a colname within \code{x})}
}
\value{
\code{mrRaw} returns an list of class
\code{mrRaw} with the
following items:
\item{df}{
A dataframe filled with adjusted estimates of a species resting
metabolic rate (RMR), torpor metabolic rate minimum (TMRmin), the
temperature at which that rate is achieved (Ttormin), and average mass of
sampled individuals.}
\item{spEM}{
A list of the \code{\link[mixtools]{spEMsymloc}} items created
which can be plotted or examined individually.}
}
\description{
\code{mrRaw} Clean the input of raw metabolic trials and return semi-
parametric metaboic estimates.
}
\details{
\code{mrRaw} uses the \code{\link[mixtools]{spEMsymloc}} method
to preform semi-parimetric fitting of data using the first and third
quantiles as priors for stocastic process.
}
