% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoursToX_fun.R
\name{prop.to.months}
\alias{prop.to.months}
\title{\code{prop.to.months}}
\usage{
prop.to.months(x)
}
\arguments{
\item{x}{precent proportion of the year}
}
\description{
\code{prop.to.months}
}
\details{
general functions for consistant exchange between time units
}
