% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleFungalGrowth_fun.R
\name{scaleFungalGrowth}
\alias{scaleFungalGrowth}
\title{Scale fungal growth rate with relative humidity.}
\usage{
scaleFungalGrowth(pct.rh, fung.params)
}
\arguments{
\item{pct.rh}{precent relative humidity}

\item{fung.params}{list of parameters passed through \code{\link{FungLoad}}}
}
\value{
returns scaled humidity scaled fungal growth rate
}
\description{
\code{scaleFungalGrowth} scales fungal growth rate with relatives
humidity.
}
\examples{
##ScaleFungalGrowthRate
data("fung.params")
fung.param <- fungalSelect("Verant")

rh <- 1:100

r <- scaleFungalGrowth(pct.rh = rh, fung.params = fung.param)

plot(rh, r, xlab = "Relative humidity (\%)", ylab = "Growth Rate",
     main = "scaledFungalGrowth")
}
\seealso{
\code{\link{fungalLoad}}, \code{\link{fungalSelect}},
\code{\link{fungalGrowth}}

Other Fungal Functions: \code{\link{fungalSelect}}
}
