% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleFungalGrowthRate_fun.R
\name{ScaleFungalGrowthRate}
\alias{ScaleFungalGrowthRate}
\title{Scale fungal growth rate with relative humidity.}
\usage{
ScaleFungalGrowthRate(pct.rh, fung.params)
}
\arguments{
\item{pct.rh}{precent relative humidity}

\item{fung.params}{list of parameters passed through \code{\link{FungLoad}}}
}
\value{
returns scaled humidity scaled fungal growth rate
}
\description{
\code{ScaleFungalGrowthRate} scales fungal growth rate with relative
humidity.
}
\examples{
##ScaleFungalGrowthRate
data("fung.params")
fung.param <- FungSelect("Verant")

rh <- 1:100

r <- ScaleFungalGrowthRate(pct.rh = rh, fung.params = fung.param)

plot(rh, r, xlab = "Relative humidity (\%)", ylab = "Growth Rate",
     main = "ScaledFungalGrowthRate")
}
\seealso{
\code{\link{FungLoad}}, \code{\link{FungSelect}},
\code{\link{FungalGrowthRate}}

Other Fungal Functions: \code{\link{FungSelect}}
}
