% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvDF_fun.R
\name{SurvDF}
\alias{SurvDF}
\title{Creates a dataframe from the survival raster for more perfect figure
generation.}
\usage{
SurvDF(surv.raster, dist.map = NULL, nights)
}
\arguments{
\item{surv.raster}{output from \code{\link{survivalRaster}}}

\item{dist.map}{shapefile containing the distribution of the species in
question.}

\item{nights}{a raster layer reprensenting the length of winter measured in nigths}
}
\value{
dataframe to be used internally for several of the plotting
functions
}
\description{
\code{SurvDF} make a pretty dataframe from the output from
\code{\link{survivalRaster}}
}
\note{
will be replaced in the next instance of the model so don't worry that
it seems to be completely redundant. it is kinda
}
