% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalMultiplot_fun.R
\name{survivalMultiplot}
\alias{survivalMultiplot}
\title{survivalMultiplot}
\usage{
survivalMultiplot(surv.stk, dist.map, nights, species.name, save.name)
}
\arguments{
\item{surv.stk}{raster result stack from \code{\link{survivalRaster}}}

\item{dist.map}{shapefile distribution of the species being modeled}

\item{nights}{a raster layer reprensenting the length of winter measured in nigths}

\item{species.name}{name of the species for plotting (generally a 4 letter}

\item{save.name}{optional argument. When a path is passed files will save to that
location with provided name. If left as defult 'NULL' will print to plot window}
}
\description{
\code{survivalMultiplot} Creates a strip figure containing survival plots for infected
 and uninfected species and a difference histogram
}
\details{
Strip figures are a combination of \code{\link{survivalPlotter}} and
\code{\link{survivalHistogram}} all tied together to create one neat strip conviently displaying
the resutlts for a single species in a single function
}
\seealso{
\code{\link{survivalRaster}}; \code{\link{SurvPlotter}}; \code{\link{survivalHistogram}};
\code{\link{dangerZone}}

Other Plot Functions: \code{\link{dangerZone}},
  \code{\link{survivalPlotter}},
  \code{\link{survivalRaster}}
}
