% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalPlotter_fun.R
\name{survivalPlotter}
\alias{survivalPlotter}
\title{Plots survival raster results including a nice outline of North America.}
\usage{
survivalPlotter(surv.stk, WNS, dist.map, nights, save.name = NULL, ...)
}
\arguments{
\item{surv.stk}{raster result stack from \code{\link{survivalRaster}}}

\item{WNS}{logical. Should the bats plotted be infected with WNS?}

\item{dist.map}{shapefile representing the distribution you wish to map the
results accross (generally speaking the entire distribution of the species)}

\item{nights}{a raster layer reprensenting the length of winter measured in nigths}
}
\value{
returns a map of North America in which the distribution supplied
is filled in with the results from the model run.
}
\description{
\code{survivalPlotter} plots the results from \code{\link{survivalRaster}} and
adds shapes of North America.
}
\seealso{
\code{\link{dangerZone}}; \code{\link{MapFigs}}; \code{link{survialRaster}};
\code{\link{survivalHistogram}}

Other Plot Functions: \code{\link{dangerZone}},
  \code{\link{survivalMultiplot}},
  \code{\link{survivalRaster}}
}
