% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvPlotter_fun.R
\name{SurvPlotter}
\alias{SurvPlotter}
\title{Plots survival raster results including a nice outline of North America.}
\usage{
SurvPlotter(surv.stk, WNS, dist.map, nights)
}
\arguments{
\item{surv.stk}{raster result stack from \code{\link{SurvivalRaster}}}

\item{WNS}{logical. Should bats have WNS}

\item{dist.map}{shapefile representing the distribution you wish to map the
results accross (generally speaking the entire distribution of the species)}

\item{nights}{a raster layer reprensenting the length of winter measured in nigths}
}
\value{
returns a map of North America in which the distribution supplied
is filled in with the results from the model run.
}
\description{
\code{SurvPlotter} plots the results from \code{\link{SurvivalRaster}} and
adds shapes of North America.
}
\note{
This function will be removed/ remodeled with the next version
}
\seealso{
\code{\link{DangerZone}}; \code{\link{MapFigs}}; \code{link{survialRaster}};
\code{\link{DiffHist}}

Other Plot Functions: \code{\link{DangerZone}},
  \code{\link{MapFigs}}, \code{\link{SurvivalRaster}}
}
