% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torpidEnergy_fun.R
\name{torporEnergy}
\alias{torporEnergy}
\title{Calculate energy per time at toprid body temp given ambient temperature}
\usage{
torporEnergy(Ta, bat.params, WNS, area, q = calcQ(Ta))
}
\arguments{
\item{Ta}{ambient emperature}

\item{bat.params}{list of bat parameters output from \code{\link{batLoad}}}

\item{q}{Q10 scaling value}
}
\value{
energy expended in volume \eqn{O^2}mL/h/g)
}
\description{
\code{torpidEnergry} Function calculating the enegry required for a
bout of torpor at a ambient temperature
}
\examples{
##CalcEnergyTimeTorpid Example
data("mylu.params")

e <- -10:34 #temperature vector

r <- torporEnergy(e, mylu.params)

plot(e, r, xlab = "Temp (C)", ylab = "Energy Expended",
     main = "torporEnergy")
}
\seealso{
\code{\link{ewl}},

Other Torpor Functions: \code{\link{ewl}}
}
\author{
Katie Haase
}
