% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torporTime_fun.R
\name{torporTime}
\alias{torporTime}
\title{Calculate maximum time in torpor}
\usage{
torporTime(Ta, pct.rh, areaPd, WNS, bat.params, fung.params)
}
\arguments{
\item{Ta}{ambient temperature}

\item{pct.rh}{precent humidity}

\item{areaPd}{surface area infected by fungus}

\item{WNS}{logical, if TRUE, considers the effect of Pd growth on EWL}

\item{bat.params}{parameters returned by \code{bat.params}}

\item{fung.params}{parameters returned by \code{fung.params}}
}
\description{
\code{torporTime} Calculates time or torpor bout given ambient temperature or EWL, whichever comes first
}
