% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatLoad_fun.R
\name{batLoad}
\alias{batLoad}
\title{Load static bat metabolic parameters}
\usage{
batLoad(x, species)
}
\arguments{
\item{x}{dataframe from which parameters are taken see \strong{Details}}

\item{species}{character string for selected species from file above}
}
\value{
returns list of names parameters used for metabolic calculations
}
\description{
\code{batLoad} Loads metabolic parameters for bats from dataframe.
}
\details{
Function loads the metabolic parameters that will be used accross
the entire distribution. Generally speaking it is suggested that you use
\code{\link{bat.params}} which is built in.
}
\examples{
##BatLoad Example
data("bat.params")
myyu.params <- batLoad(bat.params, species = "MYYU")
print(str(myyu.params))
}
