% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatLoad_fun.R
\name{BatLoad}
\alias{BatLoad}
\title{Load static bat metabolic parameters}
\usage{
BatLoad(x, species)
}
\arguments{
\item{x}{dataframe from which parameters are taken see \strong{Details}}

\item{species}{character string for selected species from file above}
}
\value{
returns list of names parameters used for metabolic calculations
}
\description{
\code{BatLoad} Loads metabolic parameters for bats from dataframe.
}
\details{
Function loads the metabolic parameters that will be used accross
the entire distribution. Generally speaking it is suggested that you use
\code{\link{BatDat}} which is built in.
}
\examples{
##BatLoad Example
data("bat.params")
myyu.params <- BatLoad(bat.params, species = "myyu")
print(str(myyu.params))
}
