% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcEWL_fun.R
\name{CalcEWL}
\alias{CalcEWL}
\title{Calculate Evaporative Water Loss}
\usage{
CalcEWL(Ta, pcr.rh, t, areaPd, mod.params = c(fung.params, bat.params),
  torpid = TRUE, WNS = c(TRUE, FALSE))
}
\arguments{
\item{t}{time}
}
\description{
\code{CalcEWL}
}
