% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTimeFlying_fun.R
\name{CalcTimeFlying}
\alias{CalcTimeFlying}
\title{Calculate the time spent flying}
\usage{
CalcTimeFlying(bat.params)
}
\arguments{
\item{bat.params}{see \code{\link{BatLoad}}}
}
\value{
a time
}
\description{
\code{calcTimeFlying} calculates the time spent flying as a proportion of
time in euthermic arrousals
}
\seealso{
\code{\link{CalcEnergyArousal}}, \code{\link{CalcCoolTime}},
\code{\link{CalcEnergyCool}}, \code{\link{CalcEnergyTimeEuthermic}},
\code{\link{CalcEnergyFlying}}, \code{\link{CalcTimeFlying}}

Other Arousal Functions: \code{\link{CalcArousalTime}},
  \code{\link{CalcEnergyArousal}},
  \code{\link{CalcEnergyCool}},
  \code{\link{CalcEnergyFlying}},
  \code{\link{CalcEnergyTimeEuthermic}}
}
