% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTimeFlying_fun.R
\name{CalcTimeFlying}
\alias{CalcTimeFlying}
\title{Calculate the time spent flying}
\usage{
CalcTimeFlying(pFly, bat.params)
}
\arguments{
\item{pFly}{proportion of euthermia spent flying}

\item{bat.params}{see \code{\link{BatLoad}}}
}
\value{
a time
}
\description{
\code{calcTimeFlying} calculates the time spent flying as a proportion of
time in euthermic arrousal
}
\seealso{
\code{\link{CalcEnergyArousal}}, \code{\link{CalcCoolTime}},
\code{\link{CalcEnergyCool}}, \code{\link{CalcEnergyTimeEuthermic}},
\code{\link{calcEnergyFlying}}, \code{\link{CalcTimeFlying}}

Other Arousal Functions: \code{\link{CalcArousalTime}},
  \code{\link{CalcEnergyArousal}},
  \code{\link{CalcEnergyCool}},
  \code{\link{CalcEnergyFlying}},
  \code{\link{CalcEnergyTimeEuthermic}}
}
