% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcTorporTimePd_fun.R
\name{CalcTorporTimePd}
\alias{CalcTorporTimePd}
\title{Calculate maximum time in torpor}
\usage{
CalcTorporTimePd(Ta, pct.rh, areaPd, WNS, mod.params = c(fung.params,
  bat.params))
}
\arguments{
\item{Ta}{ambient temperature}

\item{pct.rh}{precent humidity}

\item{areaPd}{surface area infected by fungus}

\item{WNS}{logical, if TRUE, considers the effect of Pd growth on EWL}

\item{mod.params}{list of parameters output from \code{\link{BatLoad}}
and \code{\link{FungLoad}}}
}
\description{
\code{CalcTorporTime} Calculates time or torpor bout given ambient temperature or EWL, whichever comes first
}
