% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetModel_fun.R
\name{DetModel}
\alias{DetModel}
\title{Deterministic Model of Fungal Growth}
\usage{
DetModel(t, y, params)
}
\arguments{
\item{t}{time}

\item{y}{dependant varriables calculated within
\code{\link{DynamicEnergyPd}}}

\item{params}{parameters accociated with \code{bat.params},
and \code{fung.params}}
}
\description{
\code{DetModel} solves differential equasions accociated with the
deterministic model of fungal growth.
}
\details{
internal function for \code{\link{DynamicEnergyPd}} t ocalculate the
proportion of time and energy spent in the various states that compose a hibernation.
}
\examples{
## DynamicEnegryPd_ex
data("bat.params")
data("fung.params")

my.params <- bat.params["myca",]
fung.ch <- FungSelect("Chaturvedi")


env.df <- BuildEnv(temp = c(1:10),
                pct.rh = c(90:100),
                range.res = 1,
                twinter = 9,
                winter.res = 4)

bingo <- DynamicEnergyPd(env.df, my.params, fung.ch)
}
\seealso{
\code{\link{DynamicEnergyPd}},  \code{\link{BuildEnv}}

Other Model Engine: \code{\link{BuildEnv}},
  \code{\link{DynamicEnergyPd}}
}
