% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRUpdate_fun.R
\name{mrUpdate}
\alias{mrUpdate}
\title{Updates build-in metabolic parameter estimates using new data.}
\usage{
mrUpdate(x, species, params = data("bat.params", package = "batwintor"))
}
\arguments{
\item{x}{object returned by \code{\link{mrRaw}}}

\item{species}{a character vector of the species to be updated. This should
be a row name from the \code{bat.params}}

\item{params}{a dataframe of metabolic parameters (generally that returned
with \code{\link{bat.params}})}
}
\description{
\code{mrUpdate} updates paremeter estimates supplied in
\code{\link{bat.params}} with estimates obtained through the use of
\code{\link{mrRaw}} for use throughout the package.
}
\details{
Generally speaking not all of the metabolic parameters documented
in \code{\link{bat.params}} are updated at once, and therefore
only those that have new estimates are updated.
}
\examples{
##MRUpdate
#TODO create stable example with some nature of fixed raw data
# data("bat.params")
# dat <- tbl_df(dat.dirty)
# names(dat) <- c("date", "bat", "species", "sex", "band",
# "mass.in", "mass.out", "Ta", "Ta.p",
#                 "flow", "FiO2", "FeO2", "VO2.ml.min", "VO2.ml.h", "VO2.ml.h.g")
# dat$Ta <- as.factor(dat$Ta)
# mr.raw <- MRFromRaw(x = dat, sp = "MYCA")
# bat.update <- MRUpdate(mr.raw, species = "M.californicus")
}
\seealso{
\code{\link{bat.params}}, \code{\link{mrRaw}},
\code{\link{batLoad}}
}
